[Rohan]Get all Data from database

+++ EmoloyeePayrollSevice/EmoloyeePayrollSevice/EmployeeModel.cs
@@ -9,15 +9,15 @@ namespace EmoloyeePayrollSevice
     public class EmployeeModel
     {
         public string Name { get; set; }
-        public double BasicPay { get; set; }
+        public float BasicPay { get; set; }
         public DateTime Startdate { get; set; }
         public char Gender { get; set; }
         public string Address { get; set; }
         public long PhoneNumber { get; set; }
         public string Department { get; set; }
-        public double Deduction { get; set; }
-        public double TaxablePay { get; set; }
-        public double IncomeTax { get; set; }
-        public double NetPay { get; set; }
+        public float Deduction { get; set; }
+        public float TaxablePay { get; set; }
+        public float IncomeTax { get; set; }
+        public float NetPay { get; set; }
     }
 }
